modified:   eval/run_utils.py 	modified:   exp_track/06_20_2026/ds_1.5b_magnitude.sh 	modified:   exp_track/06_20_2026/ds_1.5b_random.sh 	modified:   exp_track/06_20_2026/ds_1.5b_sparsegpt.sh 	modified:   exp_track/06_20_2026/ds_1.5b_wanda.sh 	modified:   prune/prune_magnitude.py 	modified:   prune/prune_wanda.py 	modified:   prune/wanda_utils.py 	modified:   scripts/deepseek_r1_distill_qwen_1.5b.sh 	modified:   scripts/deepseek_r1_distill_qwen_1.5b_multi_node.sh 	modified:   utils/entry_utils.py 	modified:   utils/model_utils.py

+++ prune/prune_magnitude.py
@@ -14,7 +14,7 @@ def compute_magnitude_scores(model, save_dir=None, prune_ops=None):
         layer_scores = {}
         for name, module in subset.items():
             print(f"collecting magnitude scores layer {layer_idx} name {name}")
-            layer_scores[name] = module.weight.detach().abs().cpu()
+            layer_scores[name] = module.weight.detach().float().abs().cpu()
 
         save_path = save_layer_scores_pkl(
             layer_idx,
